+++ changelog.md
@@ -2,6 +2,7 @@
 
 -  Do not look for orphans if waiting for the publisher to calm down
 -  Decrease time to cool down from 1 minute to 30 seconds
+- Make isNetworkError always be a boolean
 
 ### v11.0.0